# backend/src/main/java/kr/ac/knue/facultyevaluation/common/ApiResponse.java
package kr.ac.knue.facultyevaluation.common;

public record ApiResponse<T>(boolean success, T data, String message) {

    public static <T> ApiResponse<T> ok(T data) {
        return new ApiResponse<>(true, data, null);
    }

    public static <T> ApiResponse<T> ok(T data, String message) {
        return new ApiResponse<>(true, data, message);
    }
}