# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 교수업적평가시스템 공통기능 1차 API
  version: 0.1.0
servers:
- url: /api
security:
- sessionCookie: []
...
paths:
...
  /api/code-groups:
    get:
      operationId: listCodeGroups
      summary: 코드그룹 조회
      x-related-requirements:
      - REQ-035
      - REQ-038
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      security:
      - sessionCookie: []
      x-roles:
      - R09
      parameters:
      - name: groupId
        in: query
        required: false
        schema:
          type: string
      - name: groupName
        in: query
        required: false
        schema:
          type: string
      - name: managingDepartment
        in: query
        required: false
        schema:
          type: string
      - name: useYn
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      operationId: createCodeGroup
      summary: 코드그룹 등록
      x-related-requirements:
      - REQ-036
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 코드그룹 등록 요청자는 R09 write 권한을 가져야 한다
      - 코드그룹 등록 입력은 Requirement Registry REQ-036, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 코드그룹 등록 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 코드그룹 등록 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 코드그룹 등록 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 코드그룹 등록의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 코드그룹 등록 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCodeGroupRequest'
  /api/code-groups/{groupId}:
    put:
      operationId: updateCodeGroup
      summary: 코드그룹 수정
      x-related-requirements:
      - REQ-037
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 코드그룹 수정 요청자는 R09 write 권한을 가져야 한다
      - 코드그룹 수정 입력은 Requirement Registry REQ-037, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 코드그룹 수정 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 코드그룹 수정 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 코드그룹 수정 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 코드그룹 수정의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 코드그룹 수정 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCodeGroupRequest'
  /api/code-details:
    get:
      operationId: listCodeDetails
      summary: 상세코드 조회
      x-related-requirements:
      - REQ-039
      - REQ-042
      - REQ-069
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      security:
      - sessionCookie: []
      x-roles:
      - R09
      parameters:
      - name: groupId
        in: query
        required: false
        schema:
          type: string
      - name: filter
        in: query
        required: false
        schema:
          type: string
      - name: useYn
        in: query
        required: false
        schema:
          type: string
      - name: validOn
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      operationId: createCodeDetail
      summary: 상세코드 등록
      x-related-requirements:
      - REQ-040
      - REQ-069
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 상세코드 등록 요청자는 R09 write 권한을 가져야 한다
      - 상세코드 등록 입력은 Requirement Registry REQ-040, REQ-069, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 상세코드 등록 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 상세코드 등록 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 상세코드 등록 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 상세코드 등록의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 상세코드 등록 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCodeDetailRequest'
  /api/code-details/{groupId}/{codeValue}:
    put:
      operationId: updateCodeDetail
      summary: 상세코드 수정
      x-related-requirements:
      - REQ-041
      - REQ-042
      - REQ-005
      - REQ-006
      - REQ-043
      - REQ-046
      - REQ-051
      - REQ-060
      - REQ-003
      - REQ-049
      - REQ-050
      - REQ-007
      - REQ-054
      - REQ-055
      - REQ-057
      - REQ-066
      - REQ-062
      security:
      - sessionCookie: []
      x-roles:
      - R09
      x-business-rules:
      - 상세코드 수정 요청자는 R09 write 권한을 가져야 한다
      - 상세코드 수정 입력은 Requirement Registry REQ-041, REQ-042, REQ-005의 필수값과 불변 규칙을 만족해야 한다
      x-side-effects:
      - 상세코드 수정 결과를 관련 PostgreSQL table에 저장한다
      - 변경 전후 값, 처리자, 처리일시, 사유를 추적 가능한 구조에 남긴다
      x-required-tests:
      - 'happy: R09 세션과 유효 입력 -> 상세코드 수정 후 후속 GET에서 변경값이 보인다'
      - 'auth: 세션 없는 요청 -> 401 ApiError가 반환되고 상세코드 수정 대상 데이터는 변경되지 않는다'
      - 'validation: 필수 식별자 누락 -> 400 ApiError와 해당 field error가 반환된다'
      - 'business: 상세코드 수정의 source-backed 불변/기간/권한 규칙 위반 -> 400 또는 409로 차단되고 기존 값이 유지된다'
      - 'side-effect: 상세코드 수정 성공 -> updated_at과 변경 추적 구조에 처리자/일시가 기록된다'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      parameters:
      - name: groupId
        in: path
        required: true
        schema:
          type: string
      - name: codeValue
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCodeDetailRequest'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: KNUESESSION
...
  schemas:
    ApiResponse:
      type: object
      required:
      - success
      properties:
        success:
          type: boolean
          enum:
          - true
        data: {}
        message:
          type: string
          nullable: true
    ApiError:
      type: object
      required:
      - success
      - error
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            path:
              type: string
            timestamp:
              type: string
              format: date-time
...
    UseYn:
      type: string
      enum:
      - Y
      - N
...
    CreateCodeGroupRequest:
      type: object
      required:
      - groupId
      - groupName
      - useYn
      properties:
        groupId:
          type: string
        groupName:
          type: string
        description:
          type: string
          nullable: true
        managingDepartment:
          type: string
          nullable: true
        useYn:
          $ref: '#/components/schemas/UseYn'
    UpdateCodeGroupRequest:
      type: object
      required:
      - groupId
      - groupName
      - useYn
      properties:
        groupId:
          type: string
        groupName:
          type: string
        description:
          type: string
          nullable: true
        managingDepartment:
          type: string
          nullable: true
        useYn:
          $ref: '#/components/schemas/UseYn'
    CreateCodeDetailRequest:
      type: object
      required:
      - groupId
      - codeValue
      - codeName
      - sortOrder
      - useYn
      properties:
        groupId:
          type: string
        codeValue:
          type: string
        codeName:
          type: string
        parentCodeValue:
          type: string
          nullable: true
        sortOrder:
          type: integer
          minimum: 1
        extraAttributes:
          type: string
          nullable: true
        useYn:
          $ref: '#/components/schemas/UseYn'
        validFrom:
          type: string
          format: date
          nullable: true
        validTo:
          type: string
          format: date
          nullable: true
    UpdateCodeDetailRequest:
      type: object
      required:
      - groupId
      - codeValue
      - codeName
      - sortOrder
      - useYn
      properties:
        groupId:
          type: string
        codeValue:
          type: string
        codeName:
          type: string
        parentCodeValue:
          type: string
          nullable: true
        sortOrder:
          type: integer
          minimum: 1
        extraAttributes:
          type: string
          nullable: true
        useYn:
          $ref: '#/components/schemas/UseYn'
        validFrom:
          type: string
          format: date
          nullable: true
        validTo:
          type: string
          format: date
          nullable: true